pyproject.toml
[tool.poetry]
name = "checks_app"
version = "0.1.0"
description = "https://github.com/smenateam/assignments/tree/master/backend"
authors = ["vyachesalv <skavik46111@gmail.com>"]

[tool.poetry.dependencies]
python = "3.10.4"
Django = "^4.0.6"
python-dotenv = "^0.20.0"
psycopg2 = "^2.9.3"
djangorestframework = "^3.13.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
